@@ -13,7 +13,7 @@
     ],
     "require": {
         "ext-pcre": "*",
-        "php": ">=8.0",
+        "php": ">=8.1",
         "jakeasmith/http_build_url": "^1"
     },
     "require-dev": {
