@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Algo26\IdnaConvert;
 
 use Algo26\IdnaConvert\Punycode\FromPunycode;
@@ -6,15 +6,10 @@ use Algo26\IdnaConvert\TranscodeUnicode\TranscodeUnicode;
 
 class ToUnicode extends AbstractIdnaConvert implements IdnaConvertInterface
 {
-    /** @var TranscodeUnicode */
-    private $unicodeTransCoder;
-
-    /** @var FromPunycode */
-    private $punycodeEncoder;
+    private FromPunycode $punycodeEncoder;
 
     public function __construct()
     {
-        $this->unicodeTransCoder = new TranscodeUnicode();
         $this->punycodeEncoder = new FromPunycode();
     }
 
