@@ -1,12 +1,7 @@
-<?php
+<?php declare(strict_types=1);
 namespace Algo26\IdnaConvert\NamePrep;
 
 interface NamePrepInterface
 {
-    /**
-     * @param array $inputArray
-     *
-     * @return array
-     */
     public function do(array $inputArray): array;
 }
