@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 namespace Algo26\IdnaConvert\TranscodeUnicode;
 
 interface TranscodeUnicodeInterface
